+++ src/main/java/seedu/address/logic/parser/AddressBookParser.java
@@ -16,13 +16,12 @@ import seedu.address.logic.commands.DeleteTaskCommand;
 import seedu.address.logic.commands.EditCommand;
 import seedu.address.logic.commands.ExitCommand;
 import seedu.address.logic.commands.FindCommand;
+import seedu.address.logic.commands.FindTaskCommand;
 import seedu.address.logic.commands.HelpCommand;
 import seedu.address.logic.commands.ListAllCommand;
 import seedu.address.logic.commands.ListCommand;
 import seedu.address.logic.commands.ListTaskCommand;
 import seedu.address.logic.commands.MarkCommand;
-import seedu.address.logic.commands.ReviewCommand;
-import seedu.address.logic.commands.ReviewTaskCommand;
 import seedu.address.logic.commands.UnassignCommand;
 import seedu.address.logic.commands.UnmarkCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
@@ -64,11 +63,11 @@ public class AddressBookParser {
             return new AssignCommandParser().parse(arguments);
         case UnassignCommand.COMMAND_WORD:
             return new UnassignCommandParser().parse(arguments);
-        case ReviewCommand.COMMAND_WORD:
-            return new ReviewCommandParser().parse(arguments);
+        case FindCommand.COMMAND_WORD:
+            return new FindCommandParser().parse(arguments);
 
-        case ReviewTaskCommand.COMMAND_WORD:
-            return new ReviewTaskCommandParser().parse(arguments);
+        case FindTaskCommand.COMMAND_WORD:
+            return new FindTaskCommandParser().parse(arguments);
 
         case EditCommand.COMMAND_WORD:
             return new EditCommandParser().parse(arguments);
@@ -81,9 +80,6 @@ public class AddressBookParser {
 
         case ClearCommand.COMMAND_WORD:
             return new ClearCommand();
-
-        case FindCommand.COMMAND_WORD:
-            return new FindCommandParser().parse(arguments);
 
         case ListAllCommand.COMMAND_WORD:
             return new ListAllCommand();
