--- src/main/java/seedu/address/logic/parser/ReviewCommandParser.java
@@ -1,33 +1,0 @@
-package seedu.address.logic.parser;
-
-import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
-
-import java.util.Arrays;
-
-import seedu.address.logic.commands.ReviewCommand;
-import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.person.NameContainsExactKeywordsPredicate;
-
-/**
- * Parses input arguments and creates a new ReviewCommand object
- */
-public class ReviewCommandParser implements Parser<ReviewCommand> {
-    /**
-     * Parses the given {@code String} of arguments in the context of the ReviewCommand
-     * and returns an ReviewCommand object for execution.
-     * @throws ParseException if the user input does not conform the expected format
-     */
-    @Override
-    public ReviewCommand parse(String args) throws ParseException {
-        String trimmedArgs = args.trim();
-
-        if (trimmedArgs.isEmpty()) {
-            throw new ParseException(
-                    String.format(MESSAGE_INVALID_COMMAND_FORMAT, ReviewCommand.MESSAGE_USAGE));
-        }
-
-        String[] nameKeywords = trimmedArgs.split("\\s+");
-
-        return new ReviewCommand(new NameContainsExactKeywordsPredicate((Arrays.asList(nameKeywords))));
-    }
-}
