+++ src/test/java/seedu/address/logic/commands/FindCommandTest.java
@@ -2,22 +2,29 @@ package seedu.address.logic.commands;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
-import static seedu.address.commons.core.Messages.MESSAGE_PERSONS_LISTED_OVERVIEW;
+import static seedu.address.logic.commands.CommandTestUtil.assertCommandFailure;
 import static seedu.address.logic.commands.CommandTestUtil.assertCommandSuccess;
-import static seedu.address.model.util.TypicalPersons.CARL;
-import static seedu.address.model.util.TypicalPersons.ELLE;
-import static seedu.address.model.util.TypicalPersons.FIONA;
 import static seedu.address.model.util.TypicalPersons.getTypicalAddressBook;
+import static seedu.address.model.util.TypicalTasks.getTypicalTaskRepository;
 
 import java.util.Arrays;
 import java.util.Collections;
+import java.util.List;
 
 import org.junit.jupiter.api.Test;
 
+import javafx.collections.ObservableList;
+import seedu.address.commons.core.Messages;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
+import seedu.address.model.OfficeConnectModel;
+import seedu.address.model.Repository;
+import seedu.address.model.RepositoryModelManager;
 import seedu.address.model.UserPrefs;
-import seedu.address.model.person.NameContainsKeywordsPredicate;
+import seedu.address.model.mapping.AssignTask;
+import seedu.address.model.person.NameContainsExactKeywordsPredicate;
+import seedu.address.model.person.Person;
+import seedu.address.model.shared.Id;
 
 /**
  * Contains integration tests (interaction with the Model) for {@code FindCommand}.
@@ -26,13 +33,18 @@ public class FindCommandTest {
     private final Model model = new ModelManager(getTypicalAddressBook(), new UserPrefs());
     private final Model expectedModel = new ModelManager(getTypicalAddressBook(), new UserPrefs());
 
+    private final OfficeConnectModel officeConnectModel = new OfficeConnectModel(
+            new RepositoryModelManager<>(getTypicalTaskRepository()),
+            new RepositoryModelManager<>(new Repository<AssignTask>()));
+    private final OfficeConnectModel expectedOfficeConnectModel = new OfficeConnectModel(new
+            RepositoryModelManager<>(officeConnectModel.getTaskModelManager().getReadOnlyRepository()),
+            new RepositoryModelManager<>(new Repository<AssignTask>()));
     @Test
     public void equals() {
-        NameContainsKeywordsPredicate firstPredicate =
-                new NameContainsKeywordsPredicate(Collections.singletonList("first"));
-        NameContainsKeywordsPredicate secondPredicate =
-                new NameContainsKeywordsPredicate(Collections.singletonList("second"));
-
+        NameContainsExactKeywordsPredicate firstPredicate =
+                new NameContainsExactKeywordsPredicate(Collections.singletonList("first"));
+        NameContainsExactKeywordsPredicate secondPredicate =
+                new NameContainsExactKeywordsPredicate(Collections.singletonList("second"));
         FindCommand findFirstCommand = new FindCommand(firstPredicate);
         FindCommand findSecondCommand = new FindCommand(secondPredicate);
 
@@ -43,37 +55,67 @@ public class FindCommandTest {
         FindCommand findFirstCommandCopy = new FindCommand(firstPredicate);
         assertEquals(findFirstCommand, findFirstCommandCopy);
 
+        // different types -> returns false
+        assertNotEquals(1, findFirstCommand);
+
         // null -> returns false
         assertNotEquals(null, findFirstCommand);
 
-        // different person -> returns false
+        // different predicate -> returns false
         assertNotEquals(findFirstCommand, findSecondCommand);
     }
 
     @Test
-    public void execute_zeroKeywords_noPersonFound() {
-        String expectedMessage = String.format(MESSAGE_PERSONS_LISTED_OVERVIEW, 0);
-        NameContainsKeywordsPredicate predicate = preparePredicate(" ");
+    public void execute_invalidKeywords_noPersonFound() {
+        String expectedMessage = Messages.MESSAGE_INVALID_PERSON;
+        NameContainsExactKeywordsPredicate predicate = preparePredicate("invalidKeyword");
         FindCommand command = new FindCommand(predicate);
-        expectedModel.updateFilteredPersonList(predicate);
+        assertCommandFailure(command, model, expectedMessage);
+        // Does not flush out the GUI when taking in invalid keywords
+        assertEquals(expectedModel.getFilteredPersonList(), model.getFilteredPersonList());
+        assertEquals(expectedOfficeConnectModel.getTaskModelManager().getFilteredItemList(),
+                officeConnectModel.getTaskModelManager().getFilteredItemList());
+    }
+    @Test
+    public void execute_validKeywords_noTaskAssigned() {
+        String expectedMessage = String.format(FindCommand.MESSAGE_NO_TASK_ASSIGNED, "Alice Pauline");
+        NameContainsExactKeywordsPredicate predicate = preparePredicate("Alice Pauline");
+        FindCommand command = new FindCommand(predicate);
+
+        Id pId = getAssignedPersonId(predicate);
+
+        ObservableList<AssignTask> assignedTaskList = getAssignedTaskList(pId);
+        assertEquals(Collections.emptyList(), assignedTaskList);
+
+        expectedModel.updateFilteredPersonList(person -> person.getId().equals(pId));
+        expectedOfficeConnectModel.getTaskModelManager()
+                .updateFilteredItemList(assignedPerson -> assignedTaskList.stream()
+                .anyMatch(personTask -> personTask.getPersonId().equals(assignedPerson.getId())));
+        // Person list should be updated accordingly
         assertCommandSuccess(command, model, expectedMessage, expectedModel);
-        assertEquals(Collections.emptyList(), model.getFilteredPersonList());
+        // Task list expected to be empty
+        assertCommandSuccess(command, model, expectedMessage, expectedModel,
+                officeConnectModel, expectedOfficeConnectModel);
     }
 
-    @Test
-    public void execute_multipleKeywords_multiplePersonsFound() {
-        String expectedMessage = String.format(MESSAGE_PERSONS_LISTED_OVERVIEW, 3);
-        NameContainsKeywordsPredicate predicate = preparePredicate("Kurz Elle Kunz");
-        FindCommand command = new FindCommand(predicate);
-        expectedModel.updateFilteredPersonList(predicate);
-        assertCommandSuccess(command, model, expectedMessage, expectedModel);
-        assertEquals(Arrays.asList(CARL, ELLE, FIONA), model.getFilteredPersonList());
+    private Id getAssignedPersonId(NameContainsExactKeywordsPredicate predicate) {
+        List<Person> personList = model.getAddressBook()
+                .getPersonList()
+                .filtered(predicate);
+        return personList.get(0).getId();
+    }
+
+    private ObservableList<AssignTask> getAssignedTaskList(Id pId) {
+        return expectedOfficeConnectModel.getAssignTaskModelManager()
+                .getFilteredItemList()
+                .filtered(persontask -> persontask.getPersonId().equals(pId));
     }
 
     /**
-     * Parses {@code userInput} into a {@code NameContainsKeywordsPredicate}.
+     * Parses {@code userInput} into a {@code NameContainsExactKeywordsPredicate}.
      */
-    private NameContainsKeywordsPredicate preparePredicate(String userInput) {
-        return new NameContainsKeywordsPredicate(Arrays.asList(userInput.split("\\s+")));
+    private NameContainsExactKeywordsPredicate preparePredicate(String userInput) {
+        return new NameContainsExactKeywordsPredicate(Arrays.asList(userInput.split("\\s+")));
     }
+
 }
